% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/traitfill.R
\name{traitfill}
\alias{traitfill}
\title{Fill missing trait values}
\usage{
traitfill(dat, nrep, with_replacement = TRUE, names = c("n0", "n1"))
}
\arguments{
\item{dat}{A data set (see Details)}

\item{nrep}{Number of simulation runs to do.}

\item{with_replacement}{Do the sampling with or without
replacement (default is TRUE)}

\item{names}{Names for columns in \code{dat} corresponding to
state 0 and state1.}
}
\description{
Fill missing binary trait data values based on taxonomic structure.
}
\details{
This is a \emph{direct} port from richfitz/wood with nothing other
than stylistic changes.

It assumes that the taxonomy is structured into order / family /
genus.  Trait data are assumed in columns called "n0" and "n1",
unless alternative names are passed through in \code{names}.  A
column called "N" must be present that has the estimate of the
number of species within the genus.

So \code{dat} must have columns:
\itemize{
  \item{order}
  \item{family}
  \item{genus}
  \item{n1 (unless renamed by \code{names})}
  \item{n0 (unless renamed by \code{names})}
  \item{N (must be at least n0 + n1)}
}

The function returns a big list with elements:
\itemize{
  \item{genus: estimates at the level of genus}
  \item{family: estimates at the level of family}
  \item{order: estimates at the level of order}
  \item{overall: overall estimates}
}

each element is a data.frame with columns
\itemize{
  \item{n0, n1: observed numbers of species in each state,
    possibly renamed}
  \item{N: number of species at that taxonomic level}
  \item{K: number of species with known states}
  \item{mean, lower, upper: mean and lower/upper bounds on the
    estimated number of species in state 1}
  \item{p_mean, p_lower, p_upper: mean and lower/upper bounds on the
    estimated fraction of species in state 1}
}
}

